chore(rec-c1.7): clippy -D warnings cleanup across the cycle's touch surface
Cargo fmt --all (cosmetic), then fixes to make
'cargo clippy --workspace --all-targets -- -D warnings' pass:

* crates/chronos-services/src/execution_query.rs:
  removed unused ProjectionResult from the projection import.

* crates/chronos-services/src/projection.rs:
  - tests::make_event: drop unnecessary 'let ev = ...; ev'
    (clippy::let_and_return). Return the TraceEvent::new(...)
    expression directly.
  - tests::build_engine_preserves_seq_event_id_timestamp_ns:
    the sort/iter closures only need one of the three tuple
    fields at each site; use (_, _, seq) / (eid, _, _) patterns
    instead of binding unused 'ts' / 'eid' / 'seq'.

* crates/chronos-services/src/dual_truth_characterization.rs:
  the three RED characterization tests panic at the first
  'assert!(... is_ok(), ...)' before reaching many of the
  imports and helper functions they declare. Clippy sees the
  imports + helpers as unused. Add a file-level
  '#![allow(unused_imports, dead_code)]' with a doc comment
  explaining the design: the helpers and imports are required
  for the GREEN path the tests would take once the wrapper-side
  gate closes the divergence. Also drop three redundant
  '(40 + i) as u64' casts (40 + i is already u64).

* chronos-sandbox/tests/rec_c1_7_projection_restart_equivalence.rs:
  cargo fmt-only reformat of the unique_root helper.

* chronos-sandbox/tests/rec_c1_7_uat_c1_01_two_consumers.rs:
  cargo fmt-only reformat of the read_page helper and the
  three test bodies.

The three dual_truth tests remain RED by design — they exercise
the services directly, bypassing the wrapper-side gate. The
divergence is closed in production by the gate (see rec_c1_7
C1.7.3 commit).

+++ crates/chronos-services/src/dual_truth_characterization.rs
@@ -27,6 +27,18 @@
 //! Until the engine is rebuilt from the log, the two trees disagree on
 //! at least three operations.
 
+// These tests are intentionally RED on `main` by construction. They
+// panic before reaching the lines that consume several of the imports
+// above. Clippy therefore flags the imports as unused, but they ARE
+// required for the GREEN path that the test would take once the
+// divergence is closed by the wrapper-side gate. The `#[allow]` keeps
+// `-D warnings` clean without papering over the test design.
+//
+// Applied file-wide because the items below are referenced from tests
+// that abort before reaching the use sites; this is a documented
+// divergence characterization, not dead code.
+#![allow(unused_imports, dead_code)]
+
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -70,7 +82,7 @@ fn session_log_with(n: u64, dir: &std::path::Path) -> (SessionExecutionLog, Sess
 
     for i in 0..n {
         let ev = TraceEvent::new(
-            (40 + i) as u64,        // event_id (deliberately ≠ seq)
+            40 + i,                  // event_id (deliberately ≠ seq)
             10_000_500 + i * 1_000, // timestamp_ns
             1,                      // thread_id
             chronos_domain::EventType::FunctionEntry,
@@ -209,7 +221,7 @@ async fn dual_truth_engine_built_before_late_append_misses_late_records() {
     let first_snapshot: Vec<TraceEvent> = (0..3)
         .map(|i| {
             TraceEvent::new(
-                (40 + i) as u64,
+                40 + i,
                 10_000_500 + i * 1_000,
                 1,
                 chronos_domain::EventType::FunctionEntry,
@@ -229,7 +241,7 @@ async fn dual_truth_engine_built_before_late_append_misses_late_records() {
     // engine snapshot was taken.
     for i in 3..6 {
         let ev = TraceEvent::new(
-            (40 + i) as u64,
+            40 + i,
             10_000_500 + i * 1_000,
             1,
             chronos_domain::EventType::FunctionEntry,
